backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.1.0
Pillow==9.5.0